Make the scripts more user-friendly.
Both scripts use a environment variable ("DX9_LFI"). With this fix, they do no
longer crash but provide a helpful error message. Also, the correct import of
module "scoop" (used by run_parallel.py) is checked: if the import fails, a
hint is provided to the user about how to install it.