Baseline length is now a parameter of DXReader's constructor.
Since DX8/DX9 and DX9_Delta have different conventions about the names of the
file (presence/absence of the baseline length at the end of the file), this
patch mitigates the need to modify too many files to make the code working for
different data releases. The constructor of the classes DXReader and
DPCDX9Reader now accepts a parameter "baseline_length" (which is no longer a
parameter of their __call__() method) which is used globally. The test case and
the script run_lfidx9_ipy.py have been changed accordingly.